+++ lib/cloudinary_sdk.dart
@@ -2,6 +2,7 @@ library cloudinary_sdk;
 
 export 'package:cloudinary_sdk/src/cloudinary.dart';
 export 'package:cloudinary_sdk/src/models/cloudinary_image.dart';
+export 'package:cloudinary_sdk/src/models/cloudinary_upload_resource.dart';
 export 'package:cloudinary_sdk/src/models/cloudinary_delivery_type.dart';
 export 'package:cloudinary_sdk/src/models/cloudinary_resource_type.dart';
 export 'package:cloudinary_sdk/src/models/cloudinary_response.dart';